+++ example/storybook-nativewind/src/core-components/nativewind/gluestack-ui-provider/index.tsx
@@ -10,13 +10,13 @@ export function GluestackUIProvider({
 }: {
   mode?: 'light' | 'dark';
   children?: any;
+  style?: any;
 }) {
   return (
     <View
       style={[
         config[mode],
         { flex: 1, height: '100%', width: '100%' },
-        // @ts-ignore
         props.style,
       ]}
     >