+++ packages/unstyled/checkbox/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @gluestack-ui/checkbox
 
+## 0.1.30
+
+### Patch Changes
+
+- fix: ref typing for all the components
+- Updated dependencies
+  - @gluestack-ui/form-control@0.1.18
+
+## 0.1.29
+
+### Patch Changes
+
+- fix: removed radio and checkbox icons states
+
 ## 0.1.28
 
 ### Patch Changes